+++ yugioh_deckbuilder/src/main/java/at/htlleonding/yugioh/dao/PrintSetDAO.java
@@ -1,13 +1,12 @@
 package at.htlleonding.yugioh.dao;
 
-import at.htlleonding.yugioh.entities.Card;
 import at.htlleonding.yugioh.entities.PrintSet;
-
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 import javax.persistence.EntityManager;
 import javax.transaction.Transactional;
-import java.util.List;
+import javax.ws.rs.WebApplicationException;
+import javax.ws.rs.core.Response;
 
 @ApplicationScoped
 public class PrintSetDAO {
@@ -15,7 +14,13 @@ public class PrintSetDAO {
     EntityManager em;
 
     public PrintSet findById(Long id) {
-        return em.find(PrintSet.class, id);
+        PrintSet ps = em.find(PrintSet.class, id);
+
+        if (ps == null) {
+            throw new WebApplicationException(Response.Status.NOT_FOUND);
+        }
+
+        return ps;
     }
 
     @Transactional
@@ -26,6 +31,10 @@ public class PrintSetDAO {
 
     @Transactional
     public void removePrintSet(PrintSet printSet) {
+        if (printSet == null) {
+            throw new WebApplicationException(Response.Status.NOT_FOUND);
+        }
+
         em.remove(printSet);
     }
